Extract Metal interop sync helpers into _metal_interop.py
Move the Metal/PyTorch MPS synchronisation caching logic out of the
hot core file field.py into its own module (python/quadrants/lang/
_metal_interop.py). This addresses the feature-factorization CI check
which flags self-contained feature code being added to heavily-tracked
core files.

No behavioural change — field.py, matrix.py, and _ndarray.py now
import from _metal_interop instead of defining the helpers inline.
