+++ python/quadrants/lang/matrix.py
@@ -11,6 +11,7 @@ from quadrants._lib import core as qd_python_core
 from quadrants._lib.utils import qd_python_core as _qd_python_core
 from quadrants.lang import expr, impl, runtime_ops
 from quadrants.lang import ops as ops_mod
+from quadrants.lang._metal_interop import mps_sync_if_metal
 from quadrants.lang._ndarray import Ndarray, NdarrayHostAccess
 from quadrants.lang.common_ops import QuadrantsOperations
 from quadrants.lang.exception import (
@@ -23,7 +24,6 @@ from quadrants.lang.field import (
     Field,
     ScalarField,
     SNodeHostAccess,
-    _mps_sync_if_metal,
     _try_zerocopy_numpy,
     _try_zerocopy_torch,
 )
@@ -1486,7 +1486,7 @@ class MatrixField(Field):
 
         matrix_to_ext_arr(self, arr, as_vector)
         runtime_ops.sync()
-        _mps_sync_if_metal()
+        mps_sync_if_metal()
         return arr
 
     @python_scope
